requirements.txt
torch>=2.0
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
POT>=0.9
scikit-learn>=1.2
pyyaml>=6.0
